File Uploading: Extract into separate class
This follows the "Sprout Class" pattern in the WELC book
